Integrate USB VCP as Uart backend via factory pattern
Replace USBHostChannel with UsbHostUart, a Uart subclass backed by
USBHostDriver. Uses UartFactory (GenericFactory<Uart>) so YAML
'usb_host:' inside a uart section creates the correct type, matching
the Motor/MotorFactory pattern.

YAML config changes from top-level 'usb_host:' to nested:
  uart3:
    usb_host:
      baud: 1000000
  uart_channel3:
    uart_num: 3

- Add _factory_inst delegation to Uart (all virtual methods)
- Register UsbHostUart with UartFactory::InstanceBuilder
- Remove USBHostChannel, _usb_host/_usb_cdc from MachineConfig
- Revert USBCDCChannel Configurable additions
- Simplify Main.cpp USB block to NVS-only CDC
- Update example config and CMakeLists

+++ FluidNC/esp32/esp32s3/Console.cpp
@@ -18,7 +18,7 @@ class UartConsole : public UartChannel {
 public:
     UartConsole() : UartChannel(0, true) {}
     void init() override {
-        auto uart0 = new Uart(0);
+        auto uart0 = new Uart(uint32_t(0));
         uart0->begin(BAUD_RATE, UartData::Bits8, UartStop::Bits1, UartParity::None);
         UartChannel::init(uart0);
         // USB CDC/Host initialization is handled post-config in Main.cpp
